frechet_audio_distance/requirements.txt
absl-py>=0.7.0
apache-beam>=2.9.0
astor>=0.7.1
avro>=1.8.2
backports.functools-lru-cache>=1.5
backports.weakref>=1.0.post1
certifi>=2018.11.29
chardet>=3.0.4
crcmod>=1.7
cycler>=0.10.0
Cython>=0.29.3
dill>=0.2.8.2
docopt>=0.6.2
enum34>=1.1.6
fastavro>=0.21.17
funcsigs>=1.0.2
future>=0.17.1
futures>=3.0
gast>=0.2.2
grpcio>=1.18.0
h5py>=2.9.0
hdfs>=2.2.2
httplib2>=0.11.3
idna>=2.8
Keras-Applications>=1.0.7
Keras-Preprocessing>=1.0.8
kiwisolver>=1.0.1
Markdown>=3.0.1
matplotlib>=2.2.3
mock>=2.0.0
numpy>=1.16.0
oauth2client>=3.0.0
pbr>=5.1.2
Pillow>=5.4.1
protobuf>=3.6.1
pyasn1>=0.4.5
pyasn1-modules>=0.2.4
pydot>=1.2.4
pyparsing>=2.3.1
python-dateutil>=2.7.5
pytz>=2018.4
PyYAML>=3.13
requests>=2.21.0
rsa>=4.0
scipy>=1.2.0
six>=1.12.0
subprocess32>=3.5.3
tensorboard>=1.12.2
tensorflow>=1.12.0
#or
#tensorflow-gpu>=>=1.12.0
termcolor>=1.1.0
typing>=3.6.6
urllib3>=1.24.1
Werkzeug>=0.14.1
